Major improvements from yesterday in support of textarea, e.g. Issue Pigeon
Useful to check Extend Issue Pigeon data before save, or to use intended data
provided in download link.
Script behaves strange today with e10s on, but the changes I made yeterday are
all there.
TODO: Consolidate this version with my fxos JSON.parse and JSON.parse Add-on.
The fxos add-on is still waiting for a sandbox CSP bug fix.